services/api/requirements.txt
alembic>=1.13.2
fastapi>=0.115.0
httpx>=0.27.0
psycopg[binary]>=3.2.0
pydantic>=2.8.2
pydantic-settings>=2.4.0
PyJWT>=2.9.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.36
uvicorn[standard]>=0.30.6

pytest>=8.3.3
